--- a/.editorconfig
+++ b/.editorconfig
@@ -8,7 +8,6 @@ root = true
 [*]
 charset = utf-8
 end_of_line = lf
-indent_size = 2
 indent_style = space
 insert_final_newline = true
 trim_trailing_whitespace = true