--- a/Makefile
+++ b/Makefile
@@ -1,0 +1,7 @@
+test:
+	emacs --script test.el
+
+test-perf:
+	emacs --script perf.el
+
+.PHONY: test test-perf