--- a/Makefile
+++ b/Makefile
@@ -1,7 +1,10 @@
-test:
+parinferlib.elc: parinferlib.el
+	emacs --batch -f batch-byte-compile parinferlib.el
+
+test: parinferlib.elc
 	emacs --script test.el
 
-test-perf:
+test-perf: parinferlib.elc
 	emacs --script perf.el
 
 .PHONY: test test-perf