--- a/parinferlib.el
+++ b/parinferlib.el
@@ -1,6 +1,12 @@
 ;;; parinferlib.el --- a Parinfer implementation in Emacs Lisp
-;; v1.0.0
-;; https://github.com/oakmac/parinfer-elisp
+;;
+;; Author: Chris Oakman
+;; Homepage: https://github.com/oakmac/parinfer-elisp
+;; Version: 1.0.0
+;; Package-Requires: ((emacs "24.1"))
+;; Keywords: parinfer, extensions
+;;
+;;; Commentary:
 ;;
 ;; More information about Parinfer can be found here:
 ;; http://shaunlebron.github.io/parinfer/
@@ -11,6 +17,8 @@
 
 ;; NOTE: everything is namespaced under `parinferlib` with the assumption that
 ;;       Emacs extensions might use `parinfer`
+
+;;; Code:
 
 ;;------------------------------------------------------------------------------
 ;; Constants / Predicates
@@ -777,4 +785,4 @@
 
 (provide 'parinferlib)
 
-;;; parinfer-lib.el ends here
+;;; parinferlib.el ends here